Implement thread activation and deactivation routines

+++ Common/PhyProtocol/PSCI/PSCI.cpp
@@ -18,8 +18,8 @@ typedef enum PSCIFunctionID : UINT32
 
 static psci::PsciState CurDriverState = psci::PSCI_DRIVER_NOT_INIT;
 
-extern "C" INT64 CallSMC(UINT64 X0, UINT64 X1, UINT64 X2, UINT64 X3);
-extern "C" INT64 CallHVC(UINT64 X0, UINT64 X1, UINT64 X2, UINT64 X3);
+extern "C" INT32 CallSMC(UINT64 X0, UINT64 X1, UINT64 X2, UINT64 X3);
+extern "C" INT32 CallHVC(UINT64 X0, UINT64 X1, UINT64 X2, UINT64 X3);
 
 void psci::PSCIInit()
 {
