+++ board/mock/arch.cpp
@@ -31,12 +31,12 @@ VOID pantheon::DisableSystemTimer()
 
 }
 
-extern "C" INT64 CallSMC(UINT64 X0, UINT64 X1, UINT64 X2, UINT64 X3)
+extern "C" INT32 CallSMC(UINT64 X0, UINT64 X1, UINT64 X2, UINT64 X3)
 {
 	return 0;
 }
 
-extern "C" INT64 CallHVC(UINT64 X0, UINT64 X1, UINT64 X2, UINT64 X3)
+extern "C" INT32 CallHVC(UINT64 X0, UINT64 X1, UINT64 X2, UINT64 X3)
 {
 	return 0;
 }
